Created logic to check if local player has defuser which allows for bomb timer overlay to display "defusability" correctly.

+++ src/main/kotlin/com/charlatano/game/entity/Bomb.kt
@@ -34,7 +34,7 @@ typealias Bomb = Long
 
 internal fun Bomb.defused(): Boolean = csgoEXE.boolean(this + bBombDefused)
 
-internal fun Bomb.timeLeft(): Int = (-(engineDLL.float(dwGlobalVars + 16) - csgoEXE.float(this + flC4Blow))).toInt()
+internal fun Bomb.timeLeft(): Float = -(engineDLL.float(dwGlobalVars + 16) - csgoEXE.float(this + flC4Blow))
 
 internal fun Bomb.planted() = this != -1L && !defused() && timeLeft() > 0
